reorganize cmdline
