+++ JITLoadDll/JITLoadDll.vcxproj.filters
@@ -16,6 +16,9 @@
     <Filter Include="Header Files\asmjit">
       <UniqueIdentifier>{9b19542f-3807-46fa-9108-689343703684}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Header Files\CmdParser">
+      <UniqueIdentifier>{11eeb324-bbf3-4c7a-9193-b3b462d6f07e}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="JITLoadDll.cpp">
@@ -335,8 +338,13 @@
     <ClInclude Include="PEB.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="clipp.h">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="CmdParser\clipp.h">
+      <Filter>Header Files\CmdParser</Filter>
     </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <Text Include="CmdParser\parser.hpp">
+      <Filter>Header Files\CmdParser</Filter>
+    </Text>
   </ItemGroup>
 </Project>
